gl first triangle

+++ src/render.rs
@@ -1,22 +1,22 @@
 
 pub struct Vec3 {
 
-    x : f64,
-    y : f64,
-    z : f64,
+    x : f32,
+    y : f32,
+    z : f32,
 
 }
 
 
 impl Vec3 {
 
-    pub fn create( x : f64, y : f64, z : f64) -> Vec3 {
+    pub fn create( x : f32, y : f32, z : f32) -> Vec3 {
 
         Vec3 { x : x, y : y, z : z}
 
     }
 
-    pub fn to_array(&self) -> [f64;3] {
+    pub fn to_array(&self) -> [f32;3] {
 
         [self.x, self.y, self.z]
 
@@ -31,40 +31,36 @@ pub struct Camera {
 
     pos : Vec3,
     ang : Vec3,
-    fovx : f64,
-    fovy : f64,
+    fovx : f32,
+    fovy : f32,
 
 }
 
 
 impl Camera {
 
-    pub fn create( pos : Vec3, ang : Vec3, fovx : f64, fovy : f64) -> Camera{
+    pub fn create( pos : Vec3, ang : Vec3, fovx : f32, fovy : f32) -> Camera{
 
         Camera {pos : pos, ang : ang, fovx : fovx, fovy : fovy}
 
     }
-
-
-
-
-
-
 }
+
+
 
 pub struct LightRadial {
 
     pos : Vec3,
     color : Vec3,
-    strength : f64,
-    radius : f64,
+    strength : f32,
+    radius : f32,
 
 }
 
 
 impl LightRadial {
 
-    pub fn create( pos : Vec3, color : Vec3, strength : f64, radius : f64) -> LightRadial {
+    pub fn create( pos : Vec3, color : Vec3, strength : f32, radius : f32) -> LightRadial {
 
         LightRadial {pos : pos, color : color, strength : strength, radius : radius}
